fix: delete RandomVertiacalFlip and Random180DegFlip
These functions did not performe as intended.
Flipping the image aound is easy, but flipping the keypoint is harder
since they need to be in a certain order to make sence to the net.
The FaceCrop performce good enough alone so these functions are no longer needed and are threrfore abandoned.